refactor onRunReset
