+++ src/actions/index.ts
@@ -2,5 +2,6 @@ export { default as onStartup } from './onStartup'
 export { default as onTutorialConfig } from './onTutorialConfig'
 export { default as onTutorialContinueConfig } from './onTutorialContinueConfig'
 export { default as onValidateSetup } from './onValidateSetup'
+export { default as onRunReset } from './onRunReset'
 export { default as onErrorPage } from './onErrorPage'
 export { default as onTestPass } from './onTestPass'
