+++ src/channel.ts
@@ -1,5 +1,4 @@
 import * as T from 'typings'
-import * as TT from 'typings/tutorial'
 import * as vscode from 'vscode'
 import { setupActions, solutionActions } from './actions/setupActions'
 import { COMMANDS } from './commands'
@@ -7,10 +6,6 @@ import Context from './services/context/context'
 import logger from './services/logger'
 import { openWorkspace } from './services/workspace'
 import { showOutput } from './services/testRunner/output'
-import { exec } from './services/node'
-import reset from './services/reset'
-import getLastCommitHash from './services/reset/lastHash'
-
 import * as actions from './actions'
 
 interface Channel {
@@ -88,27 +83,7 @@ class Channel implements Channel {
         vscode.commands.executeCommand(COMMANDS.RUN_TEST, action?.payload)
         return
       case 'EDITOR_RUN_RESET':
-        // reset to timeline
-        const tutorial: TT.Tutorial | null = this.context.tutorial.get()
-        const position: T.Position = this.context.position.get()
-
-        // get last pass commit
-        const hash = getLastCommitHash(position, tutorial?.levels || [])
-
-        const branch = tutorial?.config.repo.branch
-
-        if (!branch) {
-          console.error('No repo branch found for tutorial')
-          return
-        }
-
-        // load timeline until last pass commit
-        reset({ branch, hash })
-
-        // if tutorial.config.reset.command, run it
-        if (tutorial?.config?.reset?.command) {
-          await exec({ command: tutorial.config.reset.command })
-        }
+        actions.onRunReset(this.context)
         return
       default:
         logger(`No match for action type: ${actionType}`)
